Create timer_module
After the fixed delay, this module measures the reaction time. Using the 1 ms clock, it counts the number of milliseconds until the user stops the timer (by releasing the button). The elapsed time is stored as a 24-bit binary number in the output 'time_elapsed'. This measurement represents the user's reaction time in milliseconds.