Create high_score_register.v
This module is responsible for tracking the best (lowest) reaction time achieved. Every time a new reaction time is captured, the module compares it to the currently stored high score. If the new reaction time is lower, the high score register updates accordingly. It uses a 24-bit register and initializes the high score to a maximum value (24'hFFFFFF) upon reset.